u32 -> int for qsort

+++ src/afl-fuzz-queue.c
@@ -1460,12 +1460,12 @@ void swap(u32 *a, u32 *b) {
 
 }
 
-u32 dist_partition(struct queue_entry **qbuf, u32 arr[], u32 low, u32 high) {
+int dist_partition(struct queue_entry **qbuf, u32 arr[], int low, int high) {
 
   double qdist_pi = qbuf[arr[high]]->total_dist;
-  u32 i = low - 1;
+  int i = low - 1;
 
-  for (u32 j = low; j <= high - 1; j++) {
+  for (int j = low; j <= high - 1; j++) {
     double qdist = qbuf[arr[j]]->total_dist;
     // Sort by descend distances. From large to small
     if (qdist > qdist_pi) {
@@ -1479,15 +1479,15 @@ u32 dist_partition(struct queue_entry **qbuf, u32 arr[], u32 low, u32 high) {
 
 }
 
-void dist_qsort(struct queue_entry **qbuf, u32 arr[], u32 low, u32 high) {
+void dist_qsort(struct queue_entry **qbuf, u32 arr[], int low, int high) {
 
-  DIST_LOG("dist_qsort(), low %u, high %u", low, high);
+  DIST_LOG("dist_qsort(), low %d, high %d", low, high);
 
   if (low < high) {
 
-    u32 pivot = dist_partition(qbuf, arr, low, high);
+    int pivot = dist_partition(qbuf, arr, low, high);
 
-    DIST_LOG("dist_qsort(), pivot %u", pivot);
+    DIST_LOG("dist_qsort(), pivot %d", pivot);
 
     dist_qsort(qbuf, arr, low, pivot - 1);
     dist_qsort(qbuf, arr, pivot + 1, high);
@@ -1561,7 +1561,7 @@ void dist_seed_prioritize(afl_state_t *afl) {
     PFATAL("dist_seed_prioritize(), fail to malloc %u to dist->prior_indices", dist->prior_len);
   for (u32 i = 0; i < dist->prior_len; ++i) dist->prior_indices[i] = i;
   DIST_LOG("dist->prior_len %u", dist->prior_len);
-  dist_qsort(afl->queue_buf, dist->prior_indices, 0, dist->prior_len - 1);
+  dist_qsort(afl->queue_buf, dist->prior_indices, 0, (int) dist->prior_len - 1);
 
   DIST_LOG("End of dist_seed_prioritize()");
   for (u32 i = 0; i < dist->prior_len; ++i) {
